pyproject.toml
[project]
name = "kaggle-5-days-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.18.0",
    "litellm>=1.79.3",
    "mcp>=1.21.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a25",
]
